# Replace css color rgb(...) by #...