* it's faster to generate the color code in using a table for components
* it's very likely a way faster to parse (when setting the color in the canvas)