@Snuffleupagus 2020-10-30
Interesting find; should the cases in https://github.com/mozilla/pdf.js/blob/master/src/display/svg.js be converted as well and possibly the `Util.makeCssRgb` method removed if becomes unused?