@calixteman 2020-10-31
I ran benchmarks with this file:
[foo.pdf](https://github.com/mozilla/pdf.js/files/5469056/foo.pdf)
I generated using a python script.
In this file, `makeHexColors` is called 406,890 times and the perf comparison before/after the change:
```
-- Grouped By browser, stat --
browser | stat         | Count | Baseline(ms) | Current(ms) | +/- |    %  | Result(P<.05)
------- | ------------ | ----- | ------------ | ----------- | --- | ----- | -------------
firefox | Overall      |    50 |         4578 |        4509 | -69 | -1.51 |        faster
firefox | Page Request |    50 |            6 |           8 |   2 | 25.81 |              
firefox | Rendering    |    50 |         4571 |        4501 | -71 | -1.54 |        faster
```