@timvandermeij 2020-10-31
Good improvement! If you can fix the comment above and re-run the benchmark with that fix applied, I'll also trigger the tests here.