@calixteman 2020-11-02
I already ran the benchmark with the template string.