@SamyCookie 2020-11-05
just for information see also this https://stackoverflow.com/a/55200387 that also approve the idea to make a pre created table for performance.